Remove lastPriceTimestamp cast

+++ contracts/implementation/LeveragedPool.sol
@@ -78,7 +78,7 @@ contract LeveragedPool is ILeveragedPool, Initializable {
         fee = initialization._fee;
         leverageAmount = PoolSwapLibrary.convertUIntToDecimal(initialization._leverageAmount);
         feeAddress = initialization._feeAddress;
-        lastPriceTimestamp = uint40(block.timestamp);
+        lastPriceTimestamp = block.timestamp;
         poolName = initialization._poolName;
         tokens[0] = initialization._longToken;
         tokens[1] = initialization._shortToken;
@@ -97,7 +97,7 @@ contract LeveragedPool is ILeveragedPool, Initializable {
      */
     function poolUpkeep(int256 _oldPrice, int256 _newPrice) external override onlyKeeper onlyUnpaused {
         require(intervalPassed(), "Update interval hasn't passed");
-        lastPriceTimestamp = uint40(block.timestamp);
+        lastPriceTimestamp = block.timestamp;
         // perform price change and update pool balances
         executePriceChange(_oldPrice, _newPrice);
         // execute pending commitments to enter and exit the pool
